refactor for fullnode sync gosshiping
